Исправление косяков)))

+++ src/main/java/com/example/team2/repository/DepartmentWorkerRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DepartmentWorkerRepository extends JpaRepository<DepartmentWorker, Integer> {
+public interface DepartmentWorkerRepository extends JpaRepository<DepartmentWorker, Long> {
     DepartmentWorker findByWorkerNameAndDepartment(String workerName, Department department);
     DepartmentWorker findById(long id);
 }
